fixes potentially non-terminating loop

+++ sdk/src/rplidar_driver.cpp
@@ -551,7 +551,7 @@ u_result RPlidarDriverSerialImpl::grabScanData(rplidar_response_measurement_node
 u_result RPlidarDriverSerialImpl::ascendScanData(rplidar_response_measurement_node_t * nodebuffer, size_t count)
 {
     float inc_origin_angle = 360.0/count;
-    size_t i = 0;
+    int i = 0;
 
     //Tune head
     for (i = 0; i < count; i++) {
